requirements.txt
astro-sdk-python[amazon,snowflake] >=1.1.0
scrapy 
scrapy-fake-useragent
scrapy-rotating-proxies
apache-airflow-providers-snowflake
snowflake-connector-python
apache-airflow-providers-common-sql
astronomer-providers[snowflake]
num2words
word2number
snowflake-connector-python[pandas]